Cleanup FileTreeView implementations
This commit consolidates the FileTreeView implementations into one
swoval based implementation and one nio based implementation. This was
easy to do because I simplified the FileTreeView interface so that its
input is just a Path and it returns Seq[Path, FileAttributes].

There is no need to use the swoval library to make a nio implementation.
Instead it is implemented in 5 lines in the FileTreeView companion. The
swoval implementation then doesn't need to handle the nio case. Instead
it just delegates to com.swoval.files.FileTreeViews.default. If there is
no native implementation for the target platform, that api falls back to
a swoval based nio implementation.

I moved the swoval implementation into the SwovalConverters file to limit
the number of places that com.swoval is referenced in io. It is now in
just two files: SwovalConverters.scala and FileTreeRepositoryImpl.scala.
