+++ io/src/main/scala/sbt/internal/nio/SwovalConverters.scala
@@ -10,13 +10,19 @@
 
 package sbt.internal.nio
 
+import java.io.IOException
+import java.nio.file.Path
+
+import com.swoval.files.FileTreeViews
 import com.swoval.functional.{ Either => SEither }
+import sbt.nio.file.{ FileAttributes, FileTreeView }
+
+import scala.collection.immutable.VectorBuilder
 
 /**
  * Utilities for converting between swoval and sbt data types.
  */
 private[nio] object SwovalConverters {
-
   implicit class RangeOps(val range: (Int, Int)) extends AnyVal {
     def toSwovalDepth: Int = range._2 match {
       case Int.MaxValue => Int.MaxValue
@@ -29,5 +35,23 @@ private[nio] object SwovalConverters {
         Left(com.swoval.functional.Either.leftProjection(l).getValue)
       case r: com.swoval.functional.Either.Right[L, R] => Right(f(r.get()))
     }
+  }
+}
+private[sbt] object SwovalFileTreeView extends FileTreeView.Nio[FileAttributes] {
+  private[this] val view = FileTreeViews.getDefault(true)
+  override def list(path: Path): Seq[(Path, FileAttributes)] = {
+    val result = new VectorBuilder[(Path, FileAttributes)]
+    try {
+      view.list(path, 0, _ => true).forEach { typedPath =>
+        result += typedPath.getPath ->
+          FileAttributes(
+            isDirectory = typedPath.isDirectory,
+            isOther = false,
+            isRegularFile = typedPath.isFile,
+            isSymbolicLink = typedPath.isSymbolicLink
+          )
+      }
+    } catch { case _: IOException => }
+    result.result()
   }
 }
